(minor) Great Hunt for Leaked Directx Objects, and the construction of the prefab list

+++ thomas/ThomasCore/src/thomas/graphics/GUI/GUIElements.h
@@ -126,7 +126,7 @@ namespace thomas
 
 				Vector2 PixelSize()
 				{
-					return Vector2(texture->GetWidth(), texture->GetHeight());
+					return Vector2((float)texture->GetWidth(), (float)texture->GetHeight());
 				}
 			};
 		}
